Removed bugs and imporved the setPointer method
There were a few bugs with the pointers, which have been fixed by handling more edge cases in the methods that control the current pointer (next, toFirst, toLast).

Additionally, the setPointer() method has been significantly improved. It is no longer necessary to manually specify the pointer name; you only need to indicate which pointer should be applied to the element, and the method will handle the rest.