First half of java commentary
This is the first half of the java commentary.
With this commit, there is also a bug in the pom.xml, at the moment, its not possible to compile a .jar file.

Here is also a small bit of js bug fix. There was a bug, if you clicked on getContent but with an empty list.